Add Docker support for EmailSend plugin
In this update, we've added Docker support to the EmailSend plugin. This includes a new Dockerfile that describes how to build a Docker image of the plugin. Additional lines in the project file indicate that it is intended to run on Linux. Lastly, a new docker-compose file was added for easier orchestration.